pyproject.toml
[project]
name = "autonomous-researc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "flask>=3.1.2",
    "fpdf2[markdown]>=2.8.4",
    "google-search-results>=2.4.2",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "twilio>=9.7.1",
]
